Fix screenshot cropping in landscape orientations
