+++ Sources/GXUITest/Errors.swift
@@ -1,0 +1,12 @@
+//
+//  Erroes.swift
+//  
+//
+//  Created by José Echagüe on 8/14/23.
+//
+
+import Foundation
+
+internal enum GXUITestError : Error {
+	case runtimeError(String)
+}
